% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{IR.fun}
\alias{IR.fun}
\title{Calculation of IR test statistic}
\usage{
IR.fun(u1, u2, d1, d2, copula.fam)
}
\arguments{
\item{u1}{a vector, the first pseudo-observations}

\item{u2}{a vector, the second pseudo-observations}

\item{d1}{a vector of indicators whether each observation in the first response is fully observed: \code{1} indicates the observation is fully observed, and \code{0} indicates the observation is censored}

\item{d2}{a vector of indicators whether each observation in the second response is fully observed: \code{1} indicates the observation is fully observed, and \code{0} indicates the observation is censored}

\item{copula.fam}{a character indicating which one of the following copula families: "clayton", "frank", "gumbel", and "normal"}
}
\value{
a list of the following components: \code{theta.est}, the PMLE of the copula parameter, and \code{IR}, IR test statistic.
}
\description{
Given the data and copula family, calculate the IR test statistic
}
