% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRtest_BiSurvCopula.R
\name{IRtest_BiSurvCopula}
\alias{IRtest_BiSurvCopula}
\title{IR test for misspecification of copula functions in semi-parametric survival copula model for bivariate right-censored data}
\usage{
IRtest_BiSurvCopula(
  x1,
  x2,
  d1,
  d2,
  copula.fam,
  control = list(yes.boot = TRUE, nboot = 500, seed1 = 1234, same.cen = FALSE)
)
}
\arguments{
\item{x1}{a vector, the first response}

\item{x2}{a vector, the second response}

\item{d1}{a vector of indicators whether each observation in \code{x1} is fully observed: \code {1} indicates the observation is fully observed, and \code{0} indicates the observation is censored}

\item{d2}{a vector of indicators whether each observation in \code{x2} is fully observed: \code {1} indicates the observation is fully observed, and \code{0} indicates the observation is censored}

\item{copula.fam}{a character indicating which one of the following copula families: "clayton", "frank", "gumbel", and "normal"}

\item{control}{a list of the following components: \code{yes.boot}, \code{nboot}, \code{seed1}, and \code{same.cen}. \code{yes.boot} is a logical value indicating whether to implement the bootstrap procedure. \code{nboot} is the number of bootstrap samples.  \code{seed1} is the seed for generating the bootstrap samples. \code{same.cen} is a logical value indicating whether the censoring time is same for both event time.}
}
\value{
a list of the following components: \code{theta.est}, the PMLE of the copula parameter,  \code{IR}, IR test statistic, \code{IR.boot}, the bootstrapped resamples of the IR test statistics if \code{yes.boot=TRUE}, and \code{pval}, the one-sided and two-sided p-values if \code{yes.boot=TRUE}.
}
\description{

}
