% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lr.test}
\alias{lr.test}
\title{PIOS and IR test statistic}
\usage{
lr.test(data, copula.fam, yes.PIOSexact = F)
}
\arguments{
\item{data}{a data frame or a matrix with four columns, in which the first two columns are the censoring indicator variables, and the last two columns are the estimated working variables.}

\item{copula.fam}{a character indicating which one of the following copula families: "clayton", "frank", "gumbel", and "normal".}

\item{yes.PIOSexact}{a logical value indicating whether to calculate the exact PIOS test statistic, and the default value is \code{FALSE}.}
}
\value{
a list of the following components: \code{theta.est} (the PMLE of the copula parameter), \code{PIOS.exact} (exact PIOS test statistic), \code{PIOS.approx} (approximate PIOS test statistic), and \code{IR} (IR test statistic).
}
\description{
Given the data and copula family, calculate the PIOS and IR test.
}
