% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{PIOS.fun}
\alias{PIOS.fun}
\title{Calculation of PIOS test statistic}
\usage{
PIOS.fun(u1, u2, d1, d2, copula.fam, yes.exact = F)
}
\arguments{
\item{u1}{a vector, the first pseudo-observations}

\item{u2}{a vector, the second pseudo-observations}

\item{d1}{a vector of indicators whether each observation in the first response is fully observed: \code {1} indicates the observation is fully observed, and \code{0} indicates the observation is censored}

\item{d2}{a vector of indicators whether each observation in the second response is fully observed: \code {1} indicates the observation is fully observed, and \code{0} indicates the observation is censored}

\item{copula.fam}{a character indicating which one of the following copula families: "clayton", "frank", "gumbel", and "normal"}

\item{yes.exact}{a logical value indicating whether to calculate the exact test statistic; if \code{yes.exact=FALSE} (default value), the approximate test statistic is calculated.}
}
\value{
a list of the following components: \code{theta.est}, the PMLE of the copula parameter, and \code{PIOS}, PIOS exact test statistic if \code{yes.exact=TRUE} or PIOS approximate test statistic if \code{yes.exact=FALSE} (default).
}
\description{
Given the data and copula family, calculate the PIOS test statistic.
}
